automatically backed up by learn

+++ index.js
@@ -1,2 +1,2 @@
-num = parseInt('2.222',10)
+num = parseFloat('2.222',10)
 console.log(num)
